requirements.txt
streamlit
flask
requests
transformers
torch
scipy
pandas
gtts
spacy
python-dotenv